feat(rule34): redis
