+++ src/commands/booru/rule34/api/interfaces/Rule34APIAutocomplete.ts
@@ -1,0 +1,5 @@
+export default interface Rule34APIAutocomplete {
+    readonly label: string,
+    readonly value: string,
+    readonly total: number
+}
